# Optimized keccak implementation