All tests were conducted on the same PC (Ryzen 5 5600X running at fixed 4.65 GHz).

Before:
```
test_cn_fast_hash<32> (100000 calls) - OK: 1 us/call
test_cn_fast_hash<16384> (1000 calls) - OK: 164 us/call
```

After:
```
test_cn_fast_hash<32> (100000 calls) - OK: 0 us/call
test_cn_fast_hash<16384> (1000 calls) - OK: 31 us/call
```

More than 5 times speedup for cn_fast_hash.

Also noticed consistent 1-2% improvement in `test_construct_tx` results. @j-berman this should also speed up view tags #8061