@jeffro256 2022-04-16
@SChernykh Have you tried bench-marking unrolling the outer loop `for (round = 0; round < rounds; ++round)`? Since it's a larger loop, it may not provide much benefit, but it would be interesting to see.